factor out the some of the solve-specific code into distinct modules
